Added predictive modeling and preliminary results
Adjusted the color palettes and narratives.  Getting close to a final product!

Unable to find different variables to predict urban vs. rural household electricity bills.  I like Usha's suggestion that there may be another variable that is not included in the dataset (e.g., electricity rates) that might explain the difference we are seeing.